read general information about black1996
I need to report the results.
This article does not concern the relationship between weeds herbicides and
yield but estimate the positive effect of one herbicide (sethoxydim) on the
growth and yield of wheat.
It might be interesting to report the results in the report but these are not
very useful for the quantitative analysis. I may skip it.